Make listeer socket inode a 64 bit integer, not 32
Inside the kernel, the socket inode is stored as a 64-bit value. It is
also printed out as %lu in /proc/net/tcp, which means that on 64-bit
systems, the full 64-bit inode number will be printed out in that file
too.

UNFORTUNATELY, the netlink inet_diag API only exposes the listener
socket inode as a 32 bit number??? It just silently wraps it around if
you have a lot of inodes.

Because of this:
- Internally store the inode in raingutter as uint64, since we can get
  that from /proc
- Accept the wrapped-around value from netlink, since we have no choice
  there
- Mark /proc as the preferred and most correct option for raingutter
  socket stats.

+++ raingutter/netlink_socket_stats.go
@@ -183,7 +183,11 @@ func (c *RaingutterNetlinkConnection) readStatsForFamily(family int, listenerPor
 
 		if diagReplyStruct.IdiagState == TcpListen {
 			// This is the listener socket.
-			statsRet.ListenerInode = diagReplyStruct.IdiagIndoe
+			// n.b. - inodes are usually 64-bits on Linux (hence we have statsRet.ListenerInode
+			// as uint64), but _socket_ inodes are only 32 bits (that's how they come out of the
+			// netlink API). What happens if you have more than four billion sockets open in a
+			// single network namespace is.... ???
+			statsRet.ListenerInode = uint64(diagReplyStruct.IdiagIndoe)
 			statsRet.QueueSize = uint64(diagReplyStruct.IdiagRqueue)
 		} else if isActiveTcpState(diagReplyStruct.IdiagState) {
 			statsRet.ActiveWorkers++
